test(agent): keep the ACP effort tests off the Windows build
The windows-execenv job builds the whole ./pkg/agent test binary before
applying its -run filter, so an untagged test file may only reference
symbols that exist on Windows. acp_effort_test.go used
codebuddyACPSessionResult, which lives behind //go:build !windows, and
broke the compile.

Split on that line: the provider-neutral parser, annotate, apply and gate
cases stay untagged and now genuinely run everywhere, while the two cases
needing a POSIX shell stub and the one needing the CodeBuddy capture move
to acp_effort_unix_test.go. Verified with GOOS=windows/linux/darwin go vet
./... rather than by waiting on CI again.
